+++ apps/backend/src/routes/__get.py
@@ -1,0 +1,8 @@
+from fastapi import FastAPI
+
+def main(app: FastAPI):
+    @app.get("/")
+    def execute(): 
+        return "hello"
+    
+    return execute
