+++ apps/backend/src/routes/globals.py
@@ -1,0 +1,1 @@
+codes: dict[str, tuple[str, int]] = {}
